fix: restore NPC trading to game-tick (revert regression from 4f3b789da)
Shaw's commit e4fc1abb5 (Jan 12) restored NPC trading to game-tick for
consistent 1-minute interval trading. This was accidentally removed in
commit 4f3b789da (Jan 25, "cleanup unused imports and deprecated code").

This PR restores:
- MarketDecisionEngine batch trading in game-tick
- Baseline investments via NPCInvestmentManager
- Proper imports for trading services

And removes from npc-tick:
- NPC BATCH TRADING section (now handled by game-tick)
- NPC BASELINE INVESTMENTS section (now handled by game-tick)
- Unused trading-related imports

Timeline:
- Jan 9: Trading moved to npc-tick (c9cbd1d6c)
- Jan 12: Shaw restored trading to game-tick (e4fc1abb5) ← intended state
- Jan 25: Accidentally removed again (4f3b789da) ← regression
- This PR: Restores Shaw's implementation
